+++ Sources/KawarimiHenge/Views/OverrideDetailColumnView.swift
@@ -1,4 +1,5 @@
 import KawarimiCore
+import KawarimiHengeCore
 import SwiftUI
 
 private enum OverrideDetailFocusField: Hashable {